+++ src/main/java/com/jotov/skyrunrating/repository/CompetiotionRepository.java
@@ -3,5 +3,5 @@ package com.jotov.skyrunrating.repository;
 import com.jotov.skyrunrating.entity.Competition;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CompetiotionRepository  extends CrudRepository<Competition, Integer> {
+public interface CompetiotionRepository  extends CrudRepository<Competition, Long> {
 }
