+++ src/main/java/com/jotov/skyrunrating/repository/RunnerRepository.java
@@ -3,6 +3,6 @@ package com.jotov.skyrunrating.repository;
 import com.jotov.skyrunrating.entity.Runner;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RunnerRepository extends CrudRepository<Runner, Integer> {
+public interface RunnerRepository extends CrudRepository<Runner, Long> {
 
 }
